requirements.txt
# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
dnspython>=2.5.0
disposable-email-domains>=0.0.126
py3-validate-email>=1.0.5
piapy>=0.2.0
ultra-logger>=0.1.2

# Network and security
aiohttp>=3.9.0
certifi>=2024.2.0
charset-normalizer>=3.3.0
cryptography>=42.0.0
idna>=3.6
urllib3>=2.2.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
python-dateutil>=2.8.2
pytz>=2024.1

# Development and testing
pytest>=8.0.0
black>=24.1.0
flake8>=7.0.0
tox>=4.12.0

# Documentation
Sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0